app-health-back/src/@api/app-health/infrastructure-service/resolvers/app-health-get-infrastructure-services.resolver.ts
import { AppHealthGetInfrastructureServicesHandler } from '@api/app-health/infrastructure-service';
import { AppHealthInfrastructureService } from '@api/graphql';
import { QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Query, Resolver } from '@nestjs/graphql';

@Resolver()
export class AppHealthGetInfrastructureServicesResolver
{
    constructor(
        private readonly handler: AppHealthGetInfrastructureServicesHandler,
    ) {}

    @Query('appHealthGetInfrastructureServices')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<AppHealthInfrastructureService[]>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
        );
    }
}
